+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 ++++++
 * BREAKING CHANGE: `route-filter rule create`: The `--tags` parameter is no longer supported.
 * Support Autorest 3.0 based SDKs
+* BREAKING CHANGE: `vnet peering list`, `vnet subnet list`, `routetable route list` no longer support `--ids` parameter.
 * Fix issues with update commands in `express-route`, `nsg rule`, `public-ip`, `traffic manager profile` and `vnet-gateway` where some parameters erroneously had default values.
 * `network watcher`: Added `connection-monitor` commands.
 * `network watcher show-topology`: Added support to target `--vnet` and `--subnet`.