+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.26
+++++++
+* `network dns zone create/update`: Adding support for Private DNS zones.
+
 2.0.25
 ++++++
 * BREAKING CHANGE: `route-filter rule create`: The `--tags` parameter is no longer supported.
@@ -174,7 +178,7 @@ Release History
 
 * Add support for active-active VNet gateways
 * Remove nulls values from output of `network vpn-connection list/show` commands.
-* BC: Fix bug in the output of `vpn-connection create` 
+* BC: Fix bug in the output of `vpn-connection create`
 * Fix bug where '--key-length' argument of 'vpn-connection create' was not parsed correctly.
 * Fix bug in `dns zone import` where records were not imported correctly.
 * Fix bug where `traffic-manager endpoint update` did not work.